// app/src/app/models/category.enum.ts
export enum CategoryEnum {
  vegetablesAndFruits = 'vegetablesAndFruits',
  dairyProducts = 'dairyProducts',
  breadAndPastries = 'breadAndPastries',
  meatAndFish = 'meatAndFish',
  pastaCerealsAndGrains = 'pastaCerealsAndGrains',
  frozenFood = 'frozenFood',
  cansAndSauces = 'cansAndSauces',
  snacksAndSweets = 'snacksAndSweets',
  drinks = 'drinks',
  grocery = 'grocery'
}
export enum SubCategoryEnum {
  vegetables = 'vegetables',
  fruits = 'fruits',
  greens = 'greens',
  dairy = 'dairy',
  bread = 'bread',
  milk = 'milk'
}